archon/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-dotenv
logfire
httpx
supabase
openai
langgraph
pydantic-ai
pip